web/requirements.txt
Flask==0.9
PyYAML==3.10
netifaces==0.8
Flask-SocketIO==0.1.0
flufl.enum==4.0
